services/auth-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
httpx==0.25.2
pydantic==2.5.3
